requirements.txt
click==8.0.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql==0.0.3
mysqlclient==2.1.0
PyJWT==2.3.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
